// src/module/types/Skill.ts
export enum Skill {
  Acrobatics = 'acr',
  AnimalHandling = 'ani',
  Arcana = 'arc',
  Athletics = 'ath',
  Deception = 'dec',
  History = 'his',
  Insight = 'ins',
  Investigation = 'inv',
  Intimidation = 'itm',
  Medicine = 'med',
  Nature = 'nat',
  Perception = 'prc',
  Performance = 'prf',
  Persuasion = 'per',
  Religion = 'rel',
  SleightOfHand = 'slt',
  Stealth = 'ste',
  Survival = 'sur',
}

// use this function when displaying these on screen
export const SkillLabel: { [key in Skill]: string } = {
  [Skill.Acrobatics]: 'DND5E.SkillAcr',
  [Skill.AnimalHandling]: 'DND5E.SkillAni',
  [Skill.Arcana]: 'DND5E.SkillArc',
  [Skill.Athletics]: 'DND5E.SkillAth',
  [Skill.Deception]: 'DND5E.SkillDec',
  [Skill.History]: 'DND5E.SkillHis',
  [Skill.Insight]: 'DND5E.SkillIns',
  [Skill.Investigation]: 'DND5E.SkillInv',
  [Skill.Intimidation]: 'DND5E.SkillItm',
  [Skill.Medicine]: 'DND5E.SkillMed',
  [Skill.Nature]: 'DND5E.SkillNat',
  [Skill.Perception]: 'DND5E.SkillPrc',
  [Skill.Performance]: 'DND5E.SkillPrf',
  [Skill.Persuasion]: 'DND5E.SkillPer',
  [Skill.Religion]: 'DND5E.SkillRel',
  [Skill.SleightOfHand]: 'DND5E.SkillSlt',
  [Skill.Stealth]: 'DND5E.SkillSte',
  [Skill.Survival]: 'DND5E.SkillSur',
};