// src/module/types/Language.ts
export enum Language {
  Aarakocra = 'aarakocra',
  Abyssal = 'abyssal',
  Aquan = 'aquan',
  Auran = 'auran',
  Celestial = 'celestial',
  Common = 'common',
  Deep = 'deep',
  Draconic = 'draconic',
  Druidic = 'druidic',
  Dwarvish = 'dwarvish',
  Elvish = 'elvish',
  Giant = 'giant',
  Gith = 'gith',
  Gnoll = 'gnoll',
  Gnomish = 'gnomish',
  Goblin = 'goblin',
  Halfling = 'halfling',
  Ignan = 'ignan',
  Infernal = 'infernal',
  Orc = 'orc',
  Primordial = 'primordial',
  Sylvan = 'sylvan',
  Terran = 'terran',
  ThievesCant = 'cant',
  Undercommon = 'undercommon',
}

export const LanguageLabel: { [key in Language]: string } = {
  [Language.Aarakocra]: 'DND5E.LanguagesAarakocra',
  [Language.Abyssal]: 'DND5E.LanguagesAbyssal',
  [Language.Aquan]: 'DND5E.LanguagesAquan',
  [Language.Auran]: 'DND5E.LanguagesAuran',
  [Language.Celestial]: 'DND5E.LanguagesCelestial',
  [Language.Common]: 'DND5E.LanguagesCommon',
  [Language.Deep]: 'DND5E.LanguagesDeepSpeech',
  [Language.Draconic]: 'DND5E.LanguagesDraconic',
  [Language.Druidic]: 'DND5E.LanguagesDruidic',
  [Language.Dwarvish]: 'DND5E.LanguagesDwarvish',
  [Language.Elvish]: 'DND5E.LanguagesElvish',
  [Language.Giant]: 'DND5E.LanguagesGiant',
  [Language.Gith]: 'DND5E.LanguagesGith',
  [Language.Gnoll]: 'DND5E.LanguagesGnoll',
  [Language.Gnomish]: 'DND5E.LanguagesGnomish',
  [Language.Goblin]: 'DND5E.LanguagesGoblin',
  [Language.Halfling]: 'DND5E.LanguagesHalfling',
  [Language.Ignan]: 'DND5E.LanguagesIgnan',
  [Language.Infernal]: 'DND5E.LanguagesInfernal',
  [Language.Orc]: 'DND5E.LanguagesOrc',
  [Language.Primordial]: 'DND5E.LanguagesPrimordial',
  [Language.Sylvan]: 'DND5E.LanguagesSylvan',
  [Language.Terran]: 'DND5E.LanguagesTerran',
  [Language.ThievesCant]: 'DND5E.LanguagesThievesCant',
  [Language.Undercommon]: 'DND5E.LanguagesUndercommon',
};